refactor: move roles to auth folder
