+++ packages/server/src/auth/roles.decorator.ts
@@ -1,5 +1,5 @@
 import { CustomDecorator, SetMetadata } from '@nestjs/common';
-import { Role } from '../user/enum/role.enum';
+import { Role } from './enum/role.enum';
 
 export const ROLES_KEY = 'roles';
 export const Roles = (...roles: Role[]): CustomDecorator<string> => SetMetadata(ROLES_KEY, roles);
